Check

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ char c;
 float f;
 long li;
 long long lli;
-printf("Size of a char: %ld byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of an int: %ld byte(s)\n", (unsigned long)sizeof(i));
-printf("Size of a float: %ld byte(s)\n", (unsigned long)sizeof(f));
-printf("Size of a long long int: %ld byte(s)\n", (unsigned long)sizeof(lli));
-printf("Size of a long int: %ld byte(s)\n", (unsigned long)sizeof(li));
+printf("Size of a char: %ld byte(s)\n", sizeof(c));
+printf("Size of an int: %ld byte(s)\n", sizeof(i));
+printf("Size of a float: %ld byte(s)\n", sizeof(f));
+printf("Size of a long long int: %ld byte(s)\n", sizeof(lli));
+printf("Size of a long int: %ld byte(s)\n", sizeof(li));
 return (0);
 }
